stormy-mcweatherface/pyproject.toml
[project]
name = "stormy-mcweatherface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Simon Simonsson", email = "simon.simonsson@posten.no" }
]
requires-python = ">=3.11"
dependencies = [
    "backoff>=2.2.1",
    "databricks-sdk>=0.57.0",
    "httpx>=0.28.1",
    "mlflow>=3.1.0",
    "openai>=1.91.0",
]

[project.scripts]
stormy-mcweatherface = "stormy_mcweatherface:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
